Move dask ds utils to separate file
